+++ examples/editor.rs
@@ -338,7 +338,7 @@ impl Editor<'_> {
                     }
                     input => {
                         let buffer = &mut self.buffers[self.current];
-                        buffer.modified = buffer.textarea.input(input);
+                        buffer.modified |= buffer.textarea.input(input);
                     }
                 }
             }