+++ examples/tuirs_editor.rs
@@ -341,7 +341,7 @@ impl Editor<'_> {
                     }
                     input => {
                         let buffer = &mut self.buffers[self.current];
-                        buffer.modified = buffer.textarea.input(input);
+                        buffer.modified |= buffer.textarea.input(input);
                     }
                 }
             }